Applied sf patch #492522
Fix unbalanced multi columns: the TeX multicol package supports both balanced and unbalanced columns.  However, jadetex 3.11 doesn't call the functions for unbalanced columns correctly.  Since a bug in the handling of %page-balanced-columns?% was fixed in 3.11, this is now the common case - so multicolumn documents that used to work in 3.10 no longer work...