Red Hat bugzilla bug 47345
https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=47345

Long Directory path goes off the page: If there is a long directory path in a document it goes off the side of the page.  Long pathnames should be split or put on a separate line so that they don't go off the edge of the page.  Added '\setlength\emergencystretch{2in}'. In case of overfull hboxes, an extra stretchable length gets added to the end of the lines.